+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   New Feature
   
     * Fix #1066 : Add support for Kubernetes RBAC Role and Role Binding
+    * Fix #1150: Add support for Kubernetes RBAC Cluster Role and Cluster Role Binding
     * Fix #770: Added Support for CronJob
 
 #### 4.0.0